cloud_functions/ingest_function/requirements.txt
google-cloud-storage
avro-python3